Move Texture2DExtensions.Outline() into new Outline class
